+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @inventorjs/cli
 
+## 1.5.0
+
+### Minor Changes
+
+- 2c3e370: add version
+
+### Patch Changes
+
+- Updated dependencies [2c3e370]
+  - @inventorjs/core@3.9.0
+  - @inventorjs/plugin-app@1.7.0
+  - @inventorjs/plugin-plugin@1.4.0
+
 ## 1.4.0
 
 ### Minor Changes