+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inventorjs/core
 
+## 3.9.0
+
+### Minor Changes
+
+- 2c3e370: add version
+
 ## 3.8.0
 
 ### Minor Changes