+++ packages/plugin-app-react-webpack/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @inventorjs/plugin-app-react-webpack
 
+## 1.2.0
+
+### Minor Changes
+
+- 2c3e370: add version
+
+### Patch Changes
+
+- Updated dependencies [2c3e370]
+  - @inventorjs/core@3.9.0
+
 ## 1.1.0
 
 ### Minor Changes